+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - **[Breaking Change]** `tags`, `backlink`, and `mv` (when updating links to a moved note) now ignore notes that are hidden files, that match patterns in a `.ignore` file, or that match patterns in a `.gitignore` file (if running in a path that has a git repository).
 - `tags`, `backlinks`, and `mv` should now run substantially faster in some environments thanks to the introduction of concurrency.
 - **[Breaking Change]** Removed the `graph` subcommand.
+- `new -e` now uses the `EDITOR` environment variable, if available. `ZEKE_EDITOR` continues to be supported in the same way and is preferred if both `EDITOR` and `ZEKE_EDITOR` are set.
 
 # v0.5.1